Send SYSCALL_ERROR_CLOSED from send and receive calls
If only one `Arc` reference to a Rendezvous exists, calling
`send`, `receive` or `send_receive` syscalls returns an error
`SYSCALL_ERROR_CLOSED`. This prevents waiting for messages
which can't be sent or received.

Needed to fix a bug in which Rendezvous handles weren't being
dropped if control was passed to a different thread.