Add close syscall, make STDIN, STDOUT static
When a `CommHandle` is dropped, `SYSCALL_CLOSE` is called to
remove the Rendezvous handle. If there is only one other reference
to the handle, then the Rendezvous is closed and any thread waiting
is woken up and returned an error. Threads aren't stuck waiting
forever for a message that can't be sent.

If STDIN and STDOUT are `const` then copies of them can be made
and then dropped. Adding a `close` syscall which gets called when
communication handles are dropped resulted in processes dropping
their input and output handles.